Replace Double with BigDecimal and add explicit column names

+++ plan-service/src/main/java/com/gygy/planservice/dto/ContractDto.java
@@ -5,6 +5,7 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 import java.util.UUID;
+import java.math.BigDecimal;
 
 @Data
 @NoArgsConstructor
@@ -12,6 +13,6 @@ import java.util.UUID;
 public class ContractDto {
     private UUID id;
     private String type;
-    private Double discount;
+    private BigDecimal discount;
     private PlanDto plan;
 }
